refactor(finance): extract dashboard logic to utils, enforce SRP and 200-line rule
Moved date, mapping, and alert logic from v2-finance-page to dedicated utils. No functional changes. All errors resolved. Build passes.
